+++ src/commands/misc.h
@@ -1,14 +1,16 @@
 #ifndef CMDS_MISC_H
 #define CMDS_MISC_H
 
-void cmd_quit(const char *data);
-void cmd_whois(const char *data);
-void cmd_query(const char *data);
-void cmd_names(const char *data);
-void cmd_mode(const char *data);
-void cmd_resize(const char *data);
-void cmd_away(const char *data);
-void cmd_list(const char *data);
-void cmd_banlist(const char *data);
+void cmd_quit    (const char *);
+void cmd_whois   (const char *);
+void cmd_query   (const char *);
+void cmd_names   (const char *);
+void cmd_mode    (const char *);
+void cmd_resize  (const char *);
+void cmd_away    (const char *);
+void cmd_list    (const char *);
+void cmd_banlist (const char *);
+void cmd_exlist  (const char *);
+void cmd_ilist   (const char *);
 
 #endif